fastapi_app/requirements.txt
fastapi
uvicorn
pymongo
pydantic
dnspython
python-dotenv
